+++ guillotina/request.py
@@ -20,9 +20,9 @@ from zope.interface import implementer
 import asyncio
 import enum
 import json
+import orjson
 import multidict
 import time
-import ujson
 import urllib.parse
 import uuid
 
@@ -109,9 +109,9 @@ class WebSocketMsg:
     def json(self):
         try:
             if self.type == WebSocketMsg.TEXT:
-                return ujson.loads(self.msg["text"])
+                return orjson.loads(self.msg["text"])
             else:
-                return ujson.loads(self.msg["bytes"])
+                return orjson.loads(self.msg["bytes"])
         except ValueError:
             raise WebSocketJsonDecodeError()
 
@@ -239,7 +239,7 @@ class AsgiStreamReader:
             raise asyncio.IncompleteReadError(data, n)
         return data
 
-    async def read(self, n: int = -1) -> bytes:
+    async def read(self, n: int = -1) -> Union[bytes, bytearray]:
         data = self._buffer
         while (n == -1 or len(data) < n) and not self._eof:
             chunk = await self.receive()
@@ -247,10 +247,10 @@ class AsgiStreamReader:
 
         if n == -1:
             self._buffer = bytearray()
-            return bytes(data)
+            return data
         else:
             self._buffer = data[n:]
-            return bytes(data[:n])
+            return data[:n]
 
     async def receive(self) -> bytes:
         payload = await self._asgi_receive()
@@ -548,9 +548,9 @@ class Request(object):
         bytes_body = await self.read()
         return bytes_body.decode("utf-8")
 
-    async def json(self, *, loads=json.loads) -> Any:  # type: ignore
+    async def json(self, *, loads=orjson.loads) -> Any:  # type: ignore
         """Return BODY as JSON."""
-        body = await self.text()
+        body = await self.read()
         return loads(body)
 
     @property