+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ CHANGELOG
 6.0.0a17 (unreleased)
 ---------------------
 
-- Nothing changed yet.
+- Use orjson instead of json/ujson
+  [masipcat]
 
 
 6.0.0a16 (2020-03-12)